refactor: move BookingCalculator logic to common pricing module and add documentation specs
