+++ common/src/lib.rs
@@ -5,4 +5,5 @@ pub mod geocode;
 #[cfg(not(target_arch = "wasm32"))]
 pub mod http_client;
 pub mod models;
+pub mod pricing;
 pub mod reference;
